add option continueLoading, transferring logic to base-vector-layer
