+++ tests/unit/models/new-platform-flexberry-g-i-s-map-test.js
@@ -120,7 +120,6 @@ let objB = [{
 test('it exists', function(assert) {
   let model = this.subject();
 
-  // let store = this.store();
   assert.ok(!!model);
 });
